requirements.txt
urllib
mecab-python=='0.996'